requirements.txt
Django==2.2.10
django-filter==2.2.0
djangorestframework==3.11.0
Markdown==3.2.1
Pillow==7.0.0
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.0
dj-database-url
# herokuapp requirements
gunicorn==20.0.4
whitenoise
dj_database_url
psycopg2
sendgrid
# AWS S3 requirements
boto3==1.10.45
django-storages==1.8
